Add files via upload
Added code for compound option of Put on call. Noticed there is an issue with valuing the compound options at the timestep T1 which I aim to address by potentially reversing the tree of values. In this upload the Put on call class has the indexes reversed for this process.